feat:Update README.md
Created a car park free bays monitoring system.
This has a sensor that detects cars coming and going based off their number plates. Creating a dictionary of plates of the cars still in the car park. 
A display is also implemented to show available bays, location and temperature.
Unit tests for classes implemented.